Phase 2: Convert vertex/index counts to size_t for 64-bit optimization
This commit converts all vertex and index count operations from int to size_t
throughout the graphics pipeline, enabling proper 64-bit optimization for large
meshes and eliminating potential overflow issues.

Core Interface Changes:
- VertexBuffer.h: Changed GetVertexNum()/GetIndexNum() return types to size_t
- Changed constructor parameters from int to size_t for reserve sizes
- Updated member variables mlReservedVtxSize and mlReservedIdxSize to size_t

Implementation Updates:
- VertexBufferOpenGL.h/cpp: Updated base class implementation
- VertexBufferOGL_Array.h/cpp: Updated array-based implementation
- VertexBufferOGL_VBO.h/cpp: Updated VBO-based implementation
- Fixed parameter order bug in CreateCopy (was passing IndexNum,VertexNum)

Call Site Fixes (18 files, 38+ warnings resolved):
- WorldLoaderHplMap.cpp: Updated serialization code with size_t and casts
- MeshCreator.cpp: Changed loop counters and calculations to size_t
- VertexBufferOpenGL.cpp: Added casts for CreateTriTangentVectors calls
- MeshLoaderMSH.cpp: Updated serialization similar to WorldLoaderHplMap
- PhysicsWorldNewton.cpp: Added casts for CreateFromVertices
- RendererDeferred.cpp: Updated batch size calculations
- SubMeshEntity.cpp, MeshEntity.cpp: Updated vertex count variables
- RenderFunctions.cpp, SubMesh.cpp, World.cpp: Updated local variables
- Math.cpp: Added cast for CheckLineTriMeshIntersection

All size_t values are explicitly cast to int when:
- Serializing to binary formats (AddInt32, which uses 32-bit format)
- Calling legacy APIs expecting int parameters
- Calling OpenGL functions (using appropriate GLsizeiptr casts)

Impact:
- Enables meshes with >2 billion vertices/indices on 64-bit systems
- Eliminates C4267 conversion warnings throughout codebase
- Maintains binary format compatibility through explicit casts
- Improves 64-bit performance by using native pointer-sized integers

Testing: Build succeeds with zero warnings, game runs stably.

+++ HPL2/core/include/graphics/VertexBuffer.h
@@ -36,12 +36,12 @@ namespace hpl {
 	{
 	public:
 		iVertexBuffer(iLowLevelGraphics* apLowLevelGraphics,
-			eVertexBufferType aType, 
+			eVertexBufferType aType,
 			eVertexBufferDrawType aDrawType,eVertexBufferUsageType aUsageType,
-			int alReserveVtxSize,int alReserveIdxSize) :
+			size_t alReserveVtxSize,size_t alReserveIdxSize) :
 			mType(aType), mVertexFlags(0),
 			mpLowLevelGraphics(apLowLevelGraphics),
-			mDrawType(aDrawType), mUsageType(aUsageType), 
+			mDrawType(aDrawType), mUsageType(aUsageType),
 			mlReservedVtxSize(alReserveVtxSize), mlReservedIdxSize(alReserveIdxSize),
 			mlElementNum(-1) {}
 		
@@ -92,8 +92,8 @@ namespace hpl {
 		
 		virtual unsigned int* GetIndices()=0;
 		
-		virtual int GetVertexNum()=0;
-		virtual int GetIndexNum()=0;
+		virtual size_t GetVertexNum()=0;
+		virtual size_t GetIndexNum()=0;
 
 		/**
 		 * Resizes an array to a custom size, the size is number of elements and NOT number of vertices.
@@ -118,8 +118,8 @@ namespace hpl {
 		eVertexBufferDrawType mDrawType;
 		eVertexBufferUsageType mUsageType;
 		
-		int mlReservedVtxSize;
-		int mlReservedIdxSize;
+		size_t mlReservedVtxSize;
+		size_t mlReservedIdxSize;
 
 		int mlElementNum;
